+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### next
+- close the staging area when it's emptied with a verb (e.g. on `:rm`)
+
 ### v1.12.0 - 2022-05-05
 <a name="v1.12.0"></a>
 - `:stage_all_files` internal, adding to the staging area all the files verifying the current pattern. Mapped by default to ctrl-a